Corrected visual bug and crash bug
